Avoid unnecessary JSON.stringify with webpack.DefinePlugin.
Although a string will be handled correctly (as a code fragment), using
JSON.stringify only seems to be necessary when the original value is a
string: https://webpack.js.org/plugins/define-plugin/#usage

I verified with `npm run build` that this change has no effect on the
total bundle size of the production build.

Follow-up to #51.